requirements.txt
# 藥安心 MedSafe - Python Dependencies
# Core Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0

# Chemistry / Molecular Analysis
rdkit-pypi==2024.3.5

# AI / Machine Learning
numpy==1.26.4
scikit-learn==1.5.1
shap==0.46.0

# Data Processing
pandas==2.2.2
requests==2.32.3
aiohttp==3.10.5

# PDF Generation
reportlab==4.2.2

# LINE Bot SDK
line-bot-sdk==3.11.0

# Utilities
python-dotenv==1.0.1
jinja2==3.1.4

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.0
